# unfurl/server/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PatchEnvironmentBody": {
        "additionalProperties": true,
        "description": "JSON body for /delete_deployment, /update_environment, and /delete_environment.",
        "properties": {
          "branch": {
            "default": "main",
            "description": "Target branch",
            "title": "Branch",
            "type": "string"
          },
          "commit_msg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Git commit message",
            "title": "Commit Msg"
          },
          "latest_commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Latest known commit hash for optimistic concurrency checks",
            "title": "Latest Commit"
          },
          "patch": {
            "description": "List of patch operations describing the changes to apply",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Patch",
            "type": "array"
          },
          "private_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Git personal access token or password",
            "title": "Private Token"
          },
          "queueid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "If set, the Rust proxy will enqueue the request for async processing via Redis instead of proxying synchronously",
            "title": "Queueid"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Git username for pushing the commit",
            "title": "Username"
          }
        },
        "required": [
          "patch"
        ],
        "title": "PatchEnvironmentBody",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Unfurl Server API",
    "version": "1.2.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/delete_deployment": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEnvironmentBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Conflict (repository at wrong revision)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Delete a deployment",
        "tags": [
          "Project"
        ]
      }
    },
    "/delete_environment": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEnvironmentBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Conflict (repository at wrong revision)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Delete a deployment environment",
        "tags": [
          "Project"
        ]
      }
    },
...
    "/export": {
      "get": {
        "description": "Export an ensemble or service template in a JSON format suitable for the frontend. Supports 'deployment', 'blueprint', and 'environments' formats.",
        "parameters": [
          {
            "in": "query",
            "name": "auth_project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Project ID for authorization and cache key scoping",
              "title": "Auth Project"
            }
          },
          {
            "in": "query",
            "name": "latest_commit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Commit hash used to validate the cache entry",
              "title": "Latest Commit"
            }
          },
          {
            "in": "query",
            "name": "branch",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Git branch name",
              "title": "Branch"
            }
          },
          {
            "in": "query",
            "name": "pretty",
            "required": false,
            "schema": {
              "default": false,
              "description": "Pretty-print the JSON response",
              "title": "Pretty",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "username",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Git username (alternative to X-Git-Credentials header)",
              "title": "Username"
            }
          },
          {
            "in": "query",
            "name": "visibility",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "public",
                    "private"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Repository visibility",
              "title": "Visibility"
            }
          },
          {
            "in": "query",
            "name": "queueid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "If set, the Rust proxy will enqueue the request for async processing via Redis instead of proxying synchronously",
              "title": "Queueid"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "deployment",
              "description": "Export format",
              "enum": [
                "deployment",
                "blueprint",
                "environments"
              ],
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "deployment_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Path to the deployment within the project",
              "title": "Deployment Path"
            }
          },
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Environment name (used with 'environments' format)",
              "title": "Environment"
            }
          },
          {
            "in": "query",
            "name": "include_all_deployments",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include all deployment exports embedded in the response",
              "title": "Include All Deployments",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "stale",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "const": "ok",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Return any cache hit without checking if it's out of date.",
              "title": "Stale"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            },
            "description": "GraphQL-style JSON database of TOSCA objects"
          },
          "304": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Not Modified (ETag matched)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Export ensemble as JSON",
        "tags": [
          "Export"
        ]
      }
    },
...
    "/update_environment": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEnvironmentBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Conflict (repository at wrong revision)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Update a deployment environment",
        "tags": [
          "Project"
        ]
      }
    },
...
  },
...
}